I implemented the following wish from the wiki:
---

Library view: When the cursor is on an artist in the left column, the right column is empty. Only after “space” and “j”, the respective artist's songs are shown (for the marked album). Wouldn't it be better to show all songs of all albums of the respective artist if the artist (but no album) is selected? (And only afterwards, after hitting “space” and “j”, to restrict the selection to one selected album?)

---

Note the dirty changes to rbtree.h.